+++ models/points_test.go
@@ -1154,28 +1154,28 @@ func TestParsePointMaxTimestamp(t *testing.T) {
 			models.Fields{
 				"value": 1.0,
 			},
-			time.Unix(0, int64(1<<63-1))),
+			time.Unix(0, models.MaxNanoTime)),
 	)
 }
 
 func TestParsePointMinTimestamp(t *testing.T) {
-	test(t, `cpu value=1 -9223372036854775807`,
+	test(t, `cpu value=1 -9223372036854775808`,
 		NewTestPoint(
 			"cpu",
 			models.Tags{},
 			models.Fields{
 				"value": 1.0,
 			},
-			time.Unix(0, -int64(1<<63-1))),
+			time.Unix(0, models.MinNanoTime)),
 	)
 }
 
 func TestParsePointInvalidTimestamp(t *testing.T) {
-	_, err := models.ParsePointsString("cpu value=1 9223372036854775808")
+	_, err := models.ParsePointsString("cpu value=1 92233720368547758078")
 	if err == nil {
 		t.Fatalf("ParsePoints failed: %v", err)
 	}
-	_, err = models.ParsePointsString("cpu value=1 -92233720368547758078")
+	_, err = models.ParsePointsString("cpu value=1 -9223372036854775809")
 	if err == nil {
 		t.Fatalf("ParsePoints failed: %v", err)
 	}